hpml refactoring (#7047)
